refactor(openmldb_exporter): for PyPI publish
