+++ monitoring/openmldb_exporter/__init__.py
@@ -1,0 +1,8 @@
+"""
+module openmldb_exporter
+"""
+
+from importlib.metadata import version
+
+__version__ = version(__name__)
+
