+++ monitoring/openmldb_exporter/exporter.py
@@ -23,7 +23,7 @@ import sys
 import logging
 from typing import (Iterable)
 
-from openmldb_collector import (ConfigStore, Collector, TableStatusCollector, DeployQueryStatCollector,
+from openmldb_exporter.collector import (ConfigStore, Collector, TableStatusCollector, DeployQueryStatCollector,
                                 ComponentStatusCollector)
 from prometheus_client.twisted import MetricsResource
 from sqlalchemy import engine
