+++ monitoring/pyproject.toml
@@ -1,12 +1,22 @@
 [tool.poetry]
 name = "openmldb-exporter"
-version = "0.5.0"
+version = "0.5.0a0"
 description = "prometheus exporter for OpenMLDB"
 authors = ["aceforeverd <teapot@aceforeverd.com>"]
 license = "Apache-2.0"
+readme = "README.md"
+homepage = "https://openmldb.ai"
+repository = "https://github.com/4paradigm/OpenMLDB"
+documentation = "https://openmldb.ai/docs/zh/main/maintain/monitoring.html"
+keywords = ["openmldb", "prometheus"]
+classifiers = [
+  "Programming Language :: Python :: 3.8",
+  "Topic :: System :: Monitoring",
+  "Topic :: Database",
+]
 
 [tool.poetry.scripts]
-openmldb_exporter = "openmldb_exporter:main"
+openmldb-exporter = "openmldb_exporter.exporter:main"
 
 [tool.poetry.dependencies]
 python = "^3.8"
